#18 Refactor: Test Code 패키지 구조 변경
