+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/skeleton/TestRestaurant.java
@@ -1,11 +1,11 @@
 package LikeLion.TodaysLunch.skeleton;
 
-import LikeLion.TodaysLunch.domain.FoodCategory;
-import LikeLion.TodaysLunch.domain.LocationCategory;
-import LikeLion.TodaysLunch.domain.LocationTag;
-import LikeLion.TodaysLunch.domain.Member;
-import LikeLion.TodaysLunch.domain.Menu;
-import LikeLion.TodaysLunch.domain.Restaurant;
+import LikeLion.TodaysLunch.category.domain.FoodCategory;
+import LikeLion.TodaysLunch.category.domain.LocationCategory;
+import LikeLion.TodaysLunch.category.domain.LocationTag;
+import LikeLion.TodaysLunch.member.domain.Member;
+import LikeLion.TodaysLunch.menu.domain.Menu;
+import LikeLion.TodaysLunch.restaurant.domain.Restaurant;
 
 public class TestRestaurant {
   private final TestRestaurantEnviron environ;
